Add can_cast for tuple (#1132)
Apparently `tuple` was missing an implementation for `can_cast_type`. It
likely went unnoticed because `can_cast` has a quick check to return
`True` if the source and destination types are the same python object
(`a is b` check), which would work for most cases. But for other cases
you might want it to work, the implementation was missing & causing
problems. This PR adds an implementation.

Co-authored-by: Josh Bauer <josh@sematic.dev>