Initial work to enable developing Sematic with uv (#1136)
As part of #1135, we need an alternative way to manage python
dependencies besides bazel. This PR introduces the foundation required
to use [`uv`](https://docs.astral.sh/uv/) to serve this purpose.
Introducing `uv` at python3.12 (a python version currently untested via
bazel) necessitated upgrading some packages, including linting tools, so
there are changes which may not directly appear to be related to
packaging. However, all changes in this PR should not impact logic
(beyond test logic, usually changes to this are noted).

This PR is the first in an expected series of PRs.

---------

Co-authored-by: Josh <josh@sematic.dev>