Update python support (add 3.13) (#1143)
Add support for python 3.13. Not all of the dependencies we have for
examples have support for 3.13 on all platforms, so some tests need to
be skipped in that case. Also update the Sematic worker image to use
ubuntu 24.04 and python 3.12. Leaving Sematic server at 3.12, because
the Sematic EE image needs Ray, which doesn't have full support for 3.13
yet.

Testing
-------
In addition to the new CI unit tests, ran the `add` example in the cloud
with a newly pushed Sematic worker base image.