Remove internal bazel build and Circle CI (#1145)
Now that we've migrated to a new tooling stack based on uv/GitHub
Actions, we can drop bazel and CircleCI. This PR removes those things
that are inessential.

With regards to bazel, it is important to note that there are two
factors to consider:
(1) Support for bazel to manage development and maintenance of Sematic
itself
(2) Making it possible and convenient for Sematic to run pipelines for
python code & images produced with bazel

This PR drops (1) but *not* (2). However, the bazel support for examples
was also dropped. The intent being to refer to
https://github.com/sematic-ai/example_bazel as an example of how one can
create a pipeline using Sematic + bazel.

With regards to bazel integration for (1), there are a number of
accommodations made in the code for this, the most complex being support
for producing images capable of being used as Ray head and worker nodes
along with the pipeline. To ensure that these changes did not break
that, I modified our `example_bazel` repo to use a small Ray cluster and
ran it against the version of our bazel workspace from this branch. The
branch used for this testing is
[here](https://github.com/sematic-ai/example_bazel/tree/augray/ray-test)
I confirmed that the pipeline ran, spun up an ephemeral Ray cluster, and
used it before exiting.

---------

Co-authored-by: Chance An <chance@sematic.ai>
Co-authored-by: Josh <josh@sematic.dev>