Remove EE License from Sematic, restore Apache 2.0 (#1146)
Given that we've decided to de-prioritize any monetization from Sematic,
this PR restores it to a pure Apache 2.0 license. More information on
the motivation for this change can be found
[here](https://www.gnu.org/music/free-software-song.html).

This also removes references to the EE edition in the docs and removes
the "beta" designation from some associated EE features.